Api/requirements.txt
# requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
python-multipart==0.0.6
pydantic==2.5.2
aiofiles==23.2.1
python-dotenv==1.0.0
